+++ packages/effects-assets/src/index.ts
@@ -1,11 +1,4 @@
-export class Assets {
-  constructor () {
-    console.info('effects-assets');
-  }
-}
-
 export * from './effects-package';
-
 export { default as G_QUAD } from './assets/geometries/g-quad.json';
 export { default as G_CUBE } from './assets/geometries/g-cube.json';
 export { default as M_DUCK } from './assets/materials/m-duck.json';
