+++ packages/effects-core/src/components/component.ts
@@ -1,4 +1,3 @@
-import type { Deserializer } from '../deserializer';
 import { EffectsObject } from '../effects-object';
 import { removeItem } from '../utils';
 import type { VFXItem, VFXItemContent } from '../vfx-item';
