+++ packages/effects-core/src/plugins/camera/camera-controller-node.ts
@@ -1,7 +1,6 @@
 import { clamp, Euler, Quaternion, Vector3 } from '@galacean/effects-math/es/core/index';
 import type * as spec from '@galacean/effects-specification';
 import { ItemBehaviour } from '../../components';
-import type { Deserializer, SceneData } from '../../deserializer';
 import type { Engine } from '../../engine';
 import type { ValueGetter } from '../../math';
 import { createValueGetter } from '../../math';
