+++ plugin-packages/spine/src/spine-component.ts
@@ -1,4 +1,4 @@
-import type { SceneData, BoundingBoxTriangle, HitTestTriangleParams, Engine, Deserializer, Renderer } from '@galacean/effects-core';
+import type { BoundingBoxTriangle, HitTestTriangleParams, Engine, Renderer } from '@galacean/effects-core';
 import { HitTestType, PLAYER_OPTIONS_ENV_EDITOR, RendererComponent, spec, math } from '@galacean/effects-core';
 import type { AnimationStateListener, SkeletonData, Skeleton } from './core';
 import { AnimationState, AnimationStateData } from './core';
